Student controller tests moved to new folder Student
