+++ tests/Feature/Controller/Student/StudentLanguageDetailControllerTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tests\Feature\Controller;
+namespace Tests\Feature\Controller\Student;
 
 use Tests\Fixtures\Students;
 use Tests\Fixtures\Resumes;
